[HACK] Example for add
- use xnnpack examples for now
```
python3 -m examples.backend.xnnpack_examples --model_name="add" --delegate
ls ./add_u55_fp32.pte
```
- hack up headrify even more

+++ examples/models/toy_model/model.py
@@ -51,7 +51,7 @@ class AddModule(torch.nn.Module, EagerModelBase):
         return self
 
     def get_example_inputs(self):
-        return (torch.ones(1), torch.ones(1))
+        return (torch.ones(1, dtype=torch.int8), torch.ones(1, dtype=torch.int8))
 
 
 class AddMulModule(torch.nn.Module, EagerModelBase):
